// libs/controllers/dpa-tv/dtos/dpatv-list.dto.ts
export class DpaTvListDto {
  noPeserta?: string | any;
  namaPeserta?: string | any;
  statusPeserta?: string | any;
  umur?: string | any;
  masaKerja?: string | any;
  namaPerusahaan?: string | any;
  keyword?: string | any;
  limit?: string | any;
  page?: string | any;
  gaji?: string | any;
}